requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.1.0
aniso8601==7.0.0
apns2==0.1.3
asgi-redis==1.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.2
certifi==2020.6.20
cffi==1.14.5
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
contextvars==2.4
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.6.0
Deprecated==1.2.13
diff-match-patch==20200713
dj-rest-auth==2.1.11
dj-static==0.0.6
Django==3.1.5
django-appconf==1.0.4
django-celery-beat==2.2.1
django-channels-graphql-ws==0.9.1
django-ckeditor==5.9.0
django-cors-headers==3.8.0
django-countries==7.5
django-crispy-forms==1.14.0
django-easy-select2==1.5.7
django-environ==0.6.0
django-extensions==3.1.3
django-filter==2.4.0
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-imagekit==4.0.2
django-import-export==2.3.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-nested-admin==4.0.2
django-notifications-hq==1.7.0
django-otp==1.1.3
django-paypal
django-push-notifications==3.0.0
django-redis==5.2.0
django-redis-cache==3.0.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-user-language-middleware==0.0.3
et-xmlfile==1.0.1
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
googletrans==3.1.0a0
graphene==2.1.8
graphene-django==2.13.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyper==0.5.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
immutables==0.15
incremental==21.3.0
jdcal==1.4.1
jsonfield==3.1.0
kombu==5.2.4
MarkupPy==1.14
msgpack==1.0.3
msgpack-python==0.5.6
multidict==6.0.2
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.1.1
odfpy==1.4.1
onesignal-python
onesignal-sdk==2.0.0
openpyxl==3.0.5
packaging==21.3
pilkit==2.0
Pillow==8.1.0
pip-chill==1.0.1
promise==2.3
prompt-toolkit==3.0.28
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT
pyOpenSSL==22.0.0
pyparsing==3.0.7
psycopg2-binary
pytesseract==0.3.10
python-crontab==2.6.0
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
qrcode==7.3.1
redis==3.5.3
requests==2.12.1
requests-oauthlib
rfc3986==1.5.0
Rx==1.6.1
semantic-version==2.8.5
service-identity==21.1.0
setuptools-rust==0.12.1
singledispatch==3.4.0.3
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core
sqlparse==0.3.1
static3==0.7.0
swapper==1.3.0
tablib==2.0.0
toml==0.10.2
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
Unidecode==1.1.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.14.0
xlrd==1.2.0
xlwt==1.3.0
yarl==1.7.2
zope.interface==5.4.0
generic==1.0.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
django-channels
pycountry==22.3.5
django-otp-twilio
nudepy==0.5.1
django_object_actions==4.1.0
stripe
django-storages==1.11.1
boto3
